Move Tooltip to vanilla-extract
